fix(show/rate-and-review): go to login on click
This patch updates the "rate and review" section to redirect to the
login page on click instead of when the form is submitted. This results
in a much better UX as users won't write their whole review and then
have it disappear once they hit "submit".
