+++ app/utils.ts
@@ -21,14 +21,8 @@ export function safeRedirect(
   to: FormDataEntryValue | string | null | undefined,
   defaultRedirect: string = DEFAULT_REDIRECT,
 ) {
-  if (!to || typeof to !== 'string') {
-    return defaultRedirect
-  }
-
-  if (!to.startsWith('/') || to.startsWith('//')) {
-    return defaultRedirect
-  }
-
+  if (!to || typeof to !== 'string') return defaultRedirect
+  if (!to.startsWith('/') || to.startsWith('//')) return defaultRedirect
   return to
 }
